moves some methods to new model
