+++ plugins/MobileMessaging/API.php
@@ -25,6 +25,14 @@ class API extends \Piwik\Plugin\API
     const VERIFICATION_CODE_LENGTH = 5;
     const SMS_FROM = 'Matomo';
 
+    /** @var Model $model */
+    protected $model;
+
+    public function __construct(Model $model)
+    {
+        $this->model = $model;
+    }
+
     /**
      * determine if SMS API credential are available for the current user
      *
@@ -34,29 +42,8 @@ class API extends \Piwik\Plugin\API
     {
         Piwik::checkUserHasSomeViewAccess();
 
-        $credential = $this->getSMSAPICredential();
+        $credential = $this->model->getSMSAPICredential();
         return isset($credential[MobileMessaging::API_KEY_OPTION]);
-    }
-
-    private function getSMSAPICredential()
-    {
-        $settings = $this->getCredentialManagerSettings();
-
-        $credentials = isset($settings[MobileMessaging::API_KEY_OPTION]) ? $settings[MobileMessaging::API_KEY_OPTION] : null;
-
-        // fallback for older values, where api key has been stored as string value
-        if (!empty($credentials) && !is_array($credentials)) {
-            $credentials = array(
-                'apiKey' => $credentials
-            );
-        }
-
-        return array(
-            MobileMessaging::PROVIDER_OPTION =>
-                isset($settings[MobileMessaging::PROVIDER_OPTION]) ? $settings[MobileMessaging::PROVIDER_OPTION] : null,
-            MobileMessaging::API_KEY_OPTION  =>
-                $credentials,
-        );
     }
 
     /**
@@ -67,7 +54,7 @@ class API extends \Piwik\Plugin\API
     public function getSMSProvider()
     {
         $this->checkCredentialManagementRights();
-        $credential = $this->getSMSAPICredential();
+        $credential = $this->model->getSMSAPICredential();
         return $credential[MobileMessaging::PROVIDER_OPTION];
     }
 
@@ -86,12 +73,12 @@ class API extends \Piwik\Plugin\API
         $smsProviderInstance = SMSProvider::factory($provider);
         $smsProviderInstance->verifyCredential($credentials);
 
-        $settings = $this->getCredentialManagerSettings();
+        $settings = $this->model->getCredentialManagerSettings();
 
         $settings[MobileMessaging::PROVIDER_OPTION] = $provider;
         $settings[MobileMessaging::API_KEY_OPTION] = $credentials;
 
-        $this->setCredentialManagerSettings($settings);
+        $this->model->setCredentialManagerSettings($settings);
 
         return true;
     }
@@ -123,13 +110,13 @@ class API extends \Piwik\Plugin\API
             )
         );
 
-        $this->sendSMS($smsText, $phoneNumber, self::SMS_FROM);
+        $this->model->sendSMS($smsText, $phoneNumber, self::SMS_FROM);
 
-        $phoneNumbers = $this->retrievePhoneNumbers();
+        $phoneNumbers = $this->model->retrievePhoneNumbers(Piwik::getCurrentUserLogin());
         $phoneNumbers[$phoneNumber] = $verificationCode;
-        $this->savePhoneNumbers($phoneNumbers);
+        $this->model->savePhoneNumbers(Piwik::getCurrentUserLogin(), $phoneNumbers);
 
-        $this->increaseCount(MobileMessaging::PHONE_NUMBER_VALIDATION_REQUEST_COUNT_OPTION, $phoneNumber);
+        $this->model->increaseCount(Piwik::getCurrentUserLogin(), MobileMessaging::PHONE_NUMBER_VALIDATION_REQUEST_COUNT_OPTION, $phoneNumber);
 
         return true;
     }
@@ -146,33 +133,6 @@ class API extends \Piwik\Plugin\API
     }
 
     /**
-     * send a SMS
-     *
-     * @param string $content
-     * @param string $phoneNumber
-     * @param string $from
-     * @return bool true
-     * @ignore
-     */
-    public function sendSMS($content, $phoneNumber, $from)
-    {
-        Piwik::checkUserIsNotAnonymous();
-
-        $credential = $this->getSMSAPICredential();
-        $SMSProvider = SMSProvider::factory($credential[MobileMessaging::PROVIDER_OPTION]);
-        $SMSProvider->sendSMS(
-            $credential[MobileMessaging::API_KEY_OPTION],
-            $content,
-            $phoneNumber,
-            $from
-        );
-
-        $this->increaseCount(MobileMessaging::SMS_SENT_COUNT_OPTION, $phoneNumber);
-
-        return true;
-    }
-
-    /**
      * get remaining credit
      *
      * @return string remaining credit
@@ -181,7 +141,7 @@ class API extends \Piwik\Plugin\API
     {
         $this->checkCredentialManagementRights();
 
-        $credential = $this->getSMSAPICredential();
+        $credential = $this->model->getSMSAPICredential();
         $SMSProvider = SMSProvider::factory($credential[MobileMessaging::PROVIDER_OPTION]);
         return $SMSProvider->getCreditLeft(
             $credential[MobileMessaging::API_KEY_OPTION]
@@ -199,9 +159,9 @@ class API extends \Piwik\Plugin\API
     {
         Piwik::checkUserIsNotAnonymous();
 
-        $phoneNumbers = $this->retrievePhoneNumbers();
+        $phoneNumbers = $this->model->retrievePhoneNumbers(Piwik::getCurrentUserLogin());
         unset($phoneNumbers[$phoneNumber]);
-        $this->savePhoneNumbers($phoneNumbers);
+        $this->model->savePhoneNumbers(Piwik::getCurrentUserLogin(), $phoneNumbers);
 
         /**
          * Triggered after a phone number has been deleted. This event should be used to clean up any data that is
@@ -222,48 +182,6 @@ class API extends \Piwik\Plugin\API
         return true;
     }
 
-    private function retrievePhoneNumbers()
-    {
-        $settings = $this->getCurrentUserSettings();
-
-        $phoneNumbers = array();
-        if (isset($settings[MobileMessaging::PHONE_NUMBERS_OPTION])) {
-            $phoneNumbers = $settings[MobileMessaging::PHONE_NUMBERS_OPTION];
-        }
-
-        return $phoneNumbers;
-    }
-
-    private function savePhoneNumbers($phoneNumbers)
-    {
-        $settings = $this->getCurrentUserSettings();
-
-        $settings[MobileMessaging::PHONE_NUMBERS_OPTION] = $phoneNumbers;
-
-        $this->setCurrentUserSettings($settings);
-    }
-
-    private function increaseCount($option, $phoneNumber)
-    {
-        $settings = $this->getCurrentUserSettings();
-
-        $counts = array();
-        if (isset($settings[$option])) {
-            $counts = $settings[$option];
-        }
-
-        $countToUpdate = 0;
-        if (isset($counts[$phoneNumber])) {
-            $countToUpdate = $counts[$phoneNumber];
-        }
-
-        $counts[$phoneNumber] = $countToUpdate + 1;
-
-        $settings[$option] = $counts;
-
-        $this->setCurrentUserSettings($settings);
-    }
-
     /**
      * validate phone number
      *
@@ -276,65 +194,18 @@ class API extends \Piwik\Plugin\API
     {
         Piwik::checkUserIsNotAnonymous();
 
-        $phoneNumbers = $this->retrievePhoneNumbers();
+        $phoneNumbers = $this->model->retrievePhoneNumbers(Piwik::getCurrentUserLogin());
 
         if (isset($phoneNumbers[$phoneNumber])) {
             if ($verificationCode == $phoneNumbers[$phoneNumber]) {
 
                 $phoneNumbers[$phoneNumber] = null;
-                $this->savePhoneNumbers($phoneNumbers);
+                $this->model->savePhoneNumbers(Piwik::getCurrentUserLogin(), $phoneNumbers);
                 return true;
             }
         }
 
         return false;
-    }
-
-    /**
-     * get phone number list
-     *
-     * @return array $phoneNumber => $isValidated
-     * @ignore
-     */
-    public function getPhoneNumbers()
-    {
-        Piwik::checkUserIsNotAnonymous();
-
-        $rawPhoneNumbers = $this->retrievePhoneNumbers();
-
-        $phoneNumbers = array();
-        foreach ($rawPhoneNumbers as $phoneNumber => $verificationCode) {
-            $phoneNumbers[$phoneNumber] = self::isActivated($verificationCode);
-        }
-
-        return $phoneNumbers;
-    }
-
-    /**
-     * get activated phone number list
-     *
-     * @return array $phoneNumber
-     * @ignore
-     */
-    public function getActivatedPhoneNumbers()
-    {
-        Piwik::checkUserIsNotAnonymous();
-
-        $phoneNumbers = $this->retrievePhoneNumbers();
-
-        $activatedPhoneNumbers = array();
-        foreach ($phoneNumbers as $phoneNumber => $verificationCode) {
-            if (self::isActivated($verificationCode)) {
-                $activatedPhoneNumbers[] = $phoneNumber;
-            }
-        }
-
-        return $activatedPhoneNumbers;
-    }
-
-    private static function isActivated($verificationCode)
-    {
-        return $verificationCode === null;
     }
 
     /**
@@ -346,65 +217,13 @@ class API extends \Piwik\Plugin\API
     {
         $this->checkCredentialManagementRights();
 
-        $settings = $this->getCredentialManagerSettings();
+        $settings = $this->model->getCredentialManagerSettings();
 
         $settings[MobileMessaging::API_KEY_OPTION] = null;
 
-        $this->setCredentialManagerSettings($settings);
+        $this->model->setCredentialManagerSettings($settings);
 
         return true;
-    }
-
-    private function checkCredentialManagementRights()
-    {
-        $this->getDelegatedManagement() ? Piwik::checkUserIsNotAnonymous() : Piwik::checkUserHasSuperUserAccess();
-    }
-
-    private function setUserSettings($user, $settings)
-    {
-        Option::set(
-            $user . MobileMessaging::USER_SETTINGS_POSTFIX_OPTION,
-            json_encode($settings)
-        );
-    }
-
-    private function setCurrentUserSettings($settings)
-    {
-        $this->setUserSettings(Piwik::getCurrentUserLogin(), $settings);
-    }
-
-    private function setCredentialManagerSettings($settings)
-    {
-        $this->setUserSettings($this->getCredentialManagerLogin(), $settings);
-    }
-
-    private function getCredentialManagerLogin()
-    {
-        return $this->getDelegatedManagement() ? Piwik::getCurrentUserLogin() : '';
-    }
-
-    private function getUserSettings($user)
-    {
-        $optionIndex = $user . MobileMessaging::USER_SETTINGS_POSTFIX_OPTION;
-        $userSettings = Option::get($optionIndex);
-
-        if (empty($userSettings)) {
-            $userSettings = array();
-        } else {
-            $userSettings = json_decode($userSettings, true);
-        }
-
-        return $userSettings;
-    }
-
-    private function getCredentialManagerSettings()
-    {
-        return $this->getUserSettings($this->getCredentialManagerLogin());
-    }
-
-    private function getCurrentUserSettings()
-    {
-        return $this->getUserSettings(Piwik::getCurrentUserLogin());
     }
 
     /**
@@ -415,7 +234,7 @@ class API extends \Piwik\Plugin\API
     public function setDelegatedManagement($delegatedManagement)
     {
         Piwik::checkUserHasSuperUserAccess();
-        Option::set(MobileMessaging::DELEGATED_MANAGEMENT_OPTION, $delegatedManagement);
+        $this->model->setDelegatedManagement($delegatedManagement);
     }
 
     /**
@@ -426,7 +245,11 @@ class API extends \Piwik\Plugin\API
     public function getDelegatedManagement()
     {
         Piwik::checkUserHasSomeViewAccess();
-        $option = Option::get(MobileMessaging::DELEGATED_MANAGEMENT_OPTION);
-        return $option === 'true';
+        return $this->model->getDelegatedManagement();
+    }
+
+    private function checkCredentialManagementRights()
+    {
+        $this->getDelegatedManagement() ? Piwik::checkUserIsNotAnonymous() : Piwik::checkUserHasSuperUserAccess();
     }
 }
